% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exposure.R
\name{calc_t}
\alias{calc_t}
\title{Create a trend variable}
\usage{
calc_t(n, trend = "no trend")
}
\arguments{
\item{n}{A numeric value giving the number of days to simulate.}

\item{trend}{A character string that gives the trend function to use.  Options include
cos1, cos2, cos3, linear, curvilinear, cos1linear, and no trend.}
}
\value{
A numeric vector used to generate data with seasonal trends.
}
\description{
This function creates a trend variable.
}
\examples{
calc_t(5, "cos3")

}

