% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{casecross_mod}
\alias{casecross_mod}
\title{Fit a case-crossover model}
\usage{
casecross_mod(df)
}
\arguments{
\item{df}{A data frame with exposure and outcome data}
}
\value{
Summary of the estimated log relative risk for a 1-unit increase in
   exposure
}
\description{
This function fits a case-crossover model to your data.
}
\examples{

casecross_mod(df = out)

}

