% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome.R
\name{constant_baseline}
\alias{constant_baseline}
\title{Expected baseline outcomes for constant hazard rate}
\usage{
constant_baseline(n, lambda, start.date = "2000-01-01", ...)
}
\arguments{
\item{n}{A numeric vector specifying the number of days for which to simulate
data}

\item{lambda}{A numeric value specifying the mean for the expected outcomes}

\item{start.date}{A date in the format "yyyy-mm-dd" specifying the first day
for which to simulate data}
}
\value{
A data frame with the date and expected baseline value of outcomes
   for each day of simulated data
}
\description{
This is a function to compute the expected baseline outcomes when the hazard
rate for exposure is constant.
}
\examples{
constant_baseline(n = 3, lambda = 100)

}

