% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{coverage_plot}
\alias{coverage_plot}
\title{Coverage Plot}
\usage{
coverage_plot(summarystats, true_param)
}
\arguments{
\item{summarystats}{A list or data frame of summary statistics from many
repitions of a simulation.  Must include columns titled "Estimate",
"lower_ci", and "upper_ci". This could be the first object from the output of
eesim, specified by using the format "eesim_output[[1]]".}

\item{true_param}{The true value of the parameter being estimated}
}
\value{
A plot displaying the coverage for the true value of the parameter by
the confidence intervals resulting from each repitition of the simulation.
}
\description{
This plot displays the relative risk point estimates and their confidence
intervals for each repitition of the simulation in relation to the true
relative risk.  It is a visualization of the coverage of the specified
method for the relative risk. The confidence intervals which do not contain
the true relative risk appear in red.
}
\examples{
ex_sim <- eesim(n_reps = 100, n = 10, central = 100, sd = 10,
          exposure_type = "continuous", average_outcome = 20, rr = 1.05,
          model = "spline", df_year = 1)
coverage_plot(ex_sim[[1]], true_param = 1.05)

}

