% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{create_baseline}
\alias{create_baseline}
\title{Create a series of baseline outcomes}
\usage{
create_baseline(n, average_baseline = NULL, trend = "no trend", slope = 1,
  amp = 0.6, cust_base_func = NULL, ...)
}
\arguments{
\item{n}{A numeric value specifying the number of days for which to simulate
data}

\item{average_baseline}{A numeric value specifying the average outcome value}

\item{trend}{outcome_trend A character string specifying the seasonal trend in
    health outcomes.  Options are Options are:
\itemize{
  \item{"cos1"}
  \item{"cos2"}
  \item{"cos3"}
  \item{"linear"}
  \item{"curvilinear"}
  \item{"cos1linear"}
  \item{"no trend"}
  \item{"custom"}
}}

\item{slope}{slope A numeric value specifying the slope of the trend, to be used
with trend = "linear" or trend = "cos1linear".}

\item{amp}{A numeric value specifying the amplitude of the seasonal trend.
Must be between 0 and 1.}

\item{cust_base_func}{A character string specifying a user-made custom
function for baseline trend}

\item{...}{optional arguments to a custom baseline function}
}
\value{
A numeric vector of baseline outcome values
}
\description{
This function creates a time series of baseline outcome values and allows the
user to input a custom function if desired to specify outcome trend.
}
\examples{
create_baseline(n = 5, average_baseline = 22, trend = "linear")
# create_baseline(n = 5, custom_func = "custom_baseline", outcome_type = "death")

}
