% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{crossyear_mod}
\alias{crossyear_mod}
\title{Fit a cross-year model}
\usage{
crossyear_mod(df)
}
\arguments{
\item{df}{A data frame with exposure and outcome data}
}
\value{
Summary of the estimated log relative risk for a 1-unit increase in
   exposure
}
\description{
This function fits a cross-year model to your data.
}
\examples{
crossyear_mod(df)

}

