% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{custom_baseline}
\alias{custom_baseline}
\title{Pull smoothed Chicago NMMAPS mortality data}
\usage{
custom_baseline(n, average_outcome = NA, trend = NA, outcome_type = "cvd")
}
\description{
Pull smoothed Chicago NMMAPS mortality data
}
\examples{
custom_baseline(n = 5)
custom_baseline(n = 5, outcome_type = "death")
}

