% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{rep_sims}
\alias{rep_sims}
\title{Run simulations}
\usage{
rep_sims(n_sims, model, n = 5 * 365, rr = 1.01, x_type = "continuous",
  mu = 10, sd = 2, lambda = 100)
}
\arguments{
\item{n_sims}{A numeric value specifying the number of repetitions of the
simulation to run}

\item{model}{A character string specifying which model to use.  Options
include "spline_mod" and "casecross_mod"}

\item{n}{A numeric value specifying the number of days for which to simulate
data}

\item{rr}{A numeric value specifying the relative risk for each 1-unit
increase in exposure}

\item{x_type}{A character string specifying the type of exposure data.
Options are "binary" and "continuous".}

\item{mu}{A numeric value giving the average of the exposure distribution.}

\item{sd}{A numeric value giving the standard deviation of the exposure
values around their expected values.}
}
\value{
A data frame with the summaries of the estimate of the log relative
   risk for each simulation run
}
\description{
This function runs many simulations for a given dataset and model.
}
\examples{
rep_sims(n_sims = 3, model = "spline_mod")

}

