% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exposure.R
\name{season_contexp}
\alias{season_contexp}
\title{Simulate continuous exposure data with a seasonal trend}
\usage{
season_contexp(n, mu, sd, trend, start.date = "2000-01-01", ...)
}
\arguments{
\item{n}{A numeric value giving the number of days to simulate.}

\item{mu}{A numeric vector giving the average of the exposure distribution.}

\item{sd}{A numeric vector giving the standard deviation of the exposure
distribution.}

\item{trend}{A character string that gives the trend function to use. Options
include:
\itemize{
  \item{"cos1"}
  \item{"cos2"}
  \item{"cos3"}
  \item{"linear"}
  \item{"curvilinear"}
  \item{"cos1linear"}
  \item{"no trend"}
}}
}
\value{
A numeric vector with simulated exposure of length n.
}
\description{
This function simulates a time series of continuous exposure data with a
seasonal trend.
}
\examples{
season_contexp(n = 5, mu = 100, sd = 10, trend = "cos1")

}

