% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome.R
\name{seasonal_baseline}
\alias{seasonal_baseline}
\title{Expected baseline outcomes for seasonal hazard rate}
\usage{
seasonal_baseline(n, lambda, trend, start.date = "2000-01-01", ...)
}
\arguments{
\item{n}{A numeric vector specifying the number of days for which to simulate
data}

\item{lambda}{A numeric value specifying the mean for the expected outcomes}

\item{trend}{A character string specifying the desired seasonal trend}

\item{start.date}{A date in the format "yyyy-mm-dd" specifying the first day
for which to simulate data}
}
\value{
A data frame with the date and expected baseline outcomes for each
   day of simulated data
}
\description{
This function generates expected baseline outcomes when the hazard rate for
exposure is seasonal.
}
\examples{
seasonal_baseline(n = 3, lambda = 100, trend = "cos1")

}

