% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome.R
\name{sim_binout}
\alias{sim_binout}
\title{Simulate outcome data for binary exposure with a seasonal trend}
\usage{
sim_binout(n, lambda, t, exposure, rr)
}
\arguments{
\item{n}{A numeric vector specifying the number of days for which to simulate
data}

\item{lambda}{A numeric value specifying the mean outcome count per day}

\item{t}{A numeric vector for the trend variable resulting from the
\code{calc_t} function}

\item{exposure}{A numeric vector of exposure values}

\item{rr}{A numeric value specifying the relative risk}
}
\value{
A data frame with the expected outcomes and simulated outcomes for
   each day
}
\description{
This function simulates outcome counts for binary exposure with a seasonal
trend.
}
\examples{
exposure <- season_contexp(n = 5, mu = 100, sd = 10, trend = "cos1")
t <- calc_t(n = 5, trend = "no trend")
sim_binout(n = 5, lambda = 22, t = t, exposure = exposure, rr = 1.1)

}

