% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome.R
\name{sim_contout}
\alias{sim_contout}
\title{Simulate outcomes for continuous exposure with a seasonal trend}
\usage{
sim_contout(n, mu, t, exposure, rr)
}
\arguments{
\item{n}{A numeric vector specifying the number of days for which to simulate
data}

\item{mu}{A numeric value specifying the mean}

\item{t}{A numeric vector for the trend variable resulting from the
\code{calc_t} function}

\item{exposure}{A numeric vector of exposure values}

\item{rr}{A numeric value specifying the relative risk}
}
\value{
A data frame with the expected outcome and simulated outcome for each
   day of simulated data
}
\description{
This function simulates outcome counts for continuous exposure with a
seasonal trend.
}

