% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome.R
\name{sim_data}
\alias{sim_data}
\title{Simulated Exposure and Outcome Data}
\usage{
sim_data(n, rr, lambda, exposure_type = "binary", trend = "constant",
  start.date = "2000-01-01", ...)
}
\arguments{
\item{n}{A numeric value specifying the number of days for which to simulate
data}

\item{rr}{A numeric value specifying the relative risk of the outcome for each 1-unit increase in exposure}

\item{lambda}{A numeric value specifying the mean for the expected outcomes}

\item{exposure_type}{A character string specifying the type of exposure data.  Options are "binary" and "continuous".}

\item{trend}{A character string that specifies the desired trend function.
Options are:
\itemize{
  \item{"cos1"}
  \item{"cos2"}
  \item{"cos3"}
  \item{"linear"}
  \item{"curvilinear"}
  \item{"cos1linear"}
  \item{"no trend"}
  \item{"custom"}
}}

\item{start.date}{A date in the format "yyyy-mm-dd" specifying the first day
for which to simulate data}
}
\value{
A data frame with the date, exposure value, expected baseline outcome,
expected outcome, and simulated outcome for each day of simulated data.
}
\description{
This function simulates data for exposure levels and health outcomes by date.
}
\examples{
sim_data(n=10, rr = 1.1, lambda = 100, p = .5)

}

