% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{sim_random_exposure}
\alias{sim_random_exposure}
\title{Simulate random series of exposure values}
\usage{
sim_random_exposure(n, central, trend = NA, custom_func = NULL,
  exposure_type = NA, ...)
}
\description{
Simulate random series of exposure values
}
\examples{
sim_random_exposure(n = 5, central = 0.25, exposure_type = "binary")
sim_random_exposure(n = 5, central = 100, sd = 10,
                    exposure_type = "continuous")
sim_random_exposure(n = 5, central = NA, custom_func = "custom_exposure",
                    metric = "temp")

}

