% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_performance.R
\name{beta_bias}
\alias{beta_bias}
\title{Percent Bias of Estimated Coefficient}
\usage{
beta_bias(df, true_rr)
}
\arguments{
\item{df}{A data frame of replicated simulations which must include a column
titled "Estimate"}

\item{true_rr}{The true relative risk used to simulate your data}
}
\value{
The percent bias of the mean of the estimated coefficients over n
   simulations
}
\description{
This function returns the relative bias of the mean of the estimated coefficients.
}
\examples{
sims <- create_sims(n_reps = 10, n = 600, central = 100,
                    sd = 10, exposure_type = "continuous",
                    exposure_trend = "cos1",
                    exposure_amp = 0.6,
                    average_outcome = 20,
                    outcome_trend = "no trend",
                    rr = 1.01)
fits <- fit_mods(outcome = sims, model = "spline", df_year = 1)
beta_bias(fits, true_rr = 1.02)

}
