% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_performance.R
\name{beta_var}
\alias{beta_var}
\title{Standard Deviation of Estimated Coefficients}
\usage{
beta_var(df)
}
\arguments{
\item{df}{A data frame of replicated simulations which must include columns titled "est" and "se".}
}
\value{
A data frame of the variance across all values of \eqn{\hat{\beta}} and the mean variance of the $\hat{\beta}$s
}
\description{
This function gives the variance of the point estimates of $\hat{\beta}$ over the n simulations
and the mean of the variances of each $\hat{\beta}$.
}
\examples{
beta_var(df)

}

