% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exposure.R
\name{bin_t}
\alias{bin_t}
\title{Create a trend variable for binary exposure data}
\usage{
bin_t(n, p, trend = "no trend", amp = 0.3, custom_func = NULL, ...)
}
\arguments{
\item{n}{A numeric value giving the number of days to simulate.}

\item{p}{A numeric value giving the mean probability of exposure}

\item{trend}{A character string that gives the trend function to use. Options
include:
\itemize{
  \item{"cos1"}
  \item{"cos2"}
  \item{"cos3"}
  \item{"linear"}
  \item{"curvilinear"}
  \item{"cos1linear"}
  \item{"no trend"}
}}
}
\value{
A numeric vector used to generate binary exposure data with seasonal trends
}
\description{
This function creates a trend variable for binary exposure data which is centered at p and
restricts the probability of exposure between 0 and 1.
}
\examples{
bin_t(n = 5, p = .3, trend = "cos1", amp = .3)

}

