% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exposure.R
\name{binary_exposure}
\alias{binary_exposure}
\title{Simulate binary exposure data}
\usage{
binary_exposure(n, p, trend = "no trend", slope, amp,
  start.date = "2000-01-01", custom_func = NULL, ...)
}
\arguments{
\item{n}{A numeric value specifying the number of days to simulate.}

\item{p}{A numeric value giving the mean probability of exposure}

\item{trend}{A character string that gives the trend function to use. Options
are:
\itemize{
  \item{"cos1"}
  \item{"cos2"}
  \item{"cos3"}
  \item{"linear"}
  \item{"monthly"}
  \item{"no trend"}
  \item{"custom"}
}}

\item{slope}{A numeric value specifying the slope of the trend, to be used
with trend = "linear" or trend = "cos1linear".}

\item{amp}{A numeric value specifying the amplitude of the seasonal trend.
Must be between 0 and .5.}

\item{start.date}{A date of the format "yyyy-mm-dd" from which to begin
simulating daily exposures}

\item{custom_func}{A character string specifying a customized function from
which to create a trend variable.}

\item{...}{optional arguments to a custom trend function}
}
\value{
A data frame with the dates and daily exposure values from n days
}
\description{
Simulates a time series of binary exposure values with or without seasonal
trends.
}
\examples{
binary_exposure(n = 5, p = 0.1, trend = "cos1", amp = .02,
                start.date = "2001-02-01")


}
