% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exposure.R
\name{binary_exposure}
\alias{binary_exposure}
\title{Simulate binary exposure data without a seasonal trend}
\usage{
binary_exposure(n, p, start.date = "2000-01-01", ...)
}
\arguments{
\item{n}{A numeric value giving the number of days to simulate.}

\item{p}{A numeric value between 0 and 1 specifying the probability of exposure.}

\item{start.date}{A date in the format "yyyy-mm-dd" specifying the first day of simulated measurements.}
}
\value{
A data frame with the date and a 0 or 1 exposure value for each day.
}
\description{
This function generates binary exposure data without a seasonal trend.
}
\examples{
binary_exposure(n = 5, p_exp = 0.25)

}

