% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exposure.R
\name{calc_t}
\alias{calc_t}
\title{Create a trend variable}
\usage{
calc_t(n, trend = "no trend", slope = 1, amp = 0.6, custom_func = NULL,
  ...)
}
\arguments{
\item{n}{A numeric value specifying the number of days to simulate.}

\item{trend}{A character string that specifies the desired trend function.
Options are:
\itemize{
  \item{"cos1"}
  \item{"cos2"}
  \item{"cos3"}
  \item{"linear"}
  \item{"curvilinear"}
  \item{"cos1linear"}
  \item{"no trend"}
}
See the package vignette for examples of the shapes of these trends.
Default is "no trend".}

\item{slope}{A numeric value specifying the slope of the trend, to be used
with trend = "linear" or trend = "cos1linear".}

\item{amp}{A numeric value specifying the amplitude of the seasonal trend.
Must be between -1 and 1.}

\item{custom_func}{A character string specifying a customized function from
which to create a trend variable.}

\item{...}{optional arguments to a custom trend function}
}
\value{
A numeric vector used to generate data with seasonal trends.
}
\description{
Creates a trend variable for continuous exposures.
}
\examples{
calc_t(5, "cos3", amp = .5)

}
