% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{calendar_plot}
\alias{calendar_plot}
\title{Calendar Plot}
\usage{
calendar_plot(df, type = "continuous", labels = NULL,
  legend_name = "Exposure")
}
\arguments{
\item{df}{Data frame with one column for date with entries in the format
"yyyy-mm-dd" and one column for the daily values of the variable.}

\item{type}{Character string specifying whether the exposure is continuous or
discrete}

\item{labels}{Vector of character strings naming the levels of a discrete
variable}

\item{legend_name}{Character string specifying name of variable displayed}
}
\description{
This plot displays daily values of a binary or continuous variable on a
calendar via color.
}
\examples{
testdat <- sim_exposure(n = 1000, central = 0.1,
           exposure_type = "binary")
testdat$x[c(89,101,367,500,502,598,678,700,895)] <- 3
calendar_plot(testdat, type = "discrete", labels = c("no", "yes", "maybe"))

}
