% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_performance.R
\name{check_sims}
\alias{check_sims}
\title{Model Performance Assessment}
\usage{
check_sims(df, true_rr)
}
\arguments{
\item{df}{A data frame of replicated simulations which must include a column titled "est"}

\item{true_rr}{The true relative risk used to simulate your data}
}
\value{
Mean beta estimate, mean relative risk estimate, variance across betas, mean variance of
the estimates, percent bias, coverage, and power.
}
\description{
This function gives several measures of model performance.
}
\examples{
check_sims(df, true_rr = 1.02)

}

