% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exposure.R
\name{continuous_exposure}
\alias{continuous_exposure}
\title{Simulate continuous exposure data without a seasonal trend}
\usage{
continuous_exposure(n, mu, sd, start.date = "2000-01-01", ...)
}
\arguments{
\item{n}{A numeric value giving the number of days to simulate.}

\item{mu}{A numeric vector giving the average of the exposure distribution.}

\item{sd}{A numeric vector giving the standard deviation of the exposure
distribution.}

\item{start.date}{A date in the format "yyyy-mm-dd" specifying the first day of simulated measurements.}
}
\value{
A data frame with date and exposure value for each day.
}
\description{
This function simulates continuous exposure data without a seasonal trend.
}
\examples{
continuous_exposure(n = 5, mu = 10, sd = 10)

}

