% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exposure.R
\name{continuous_exposure}
\alias{continuous_exposure}
\title{Simulate continuous exposure data}
\usage{
continuous_exposure(n, mu, sd = 1, trend = "no trend", amp = 0.6,
  start.date = "2000-01-01", ...)
}
\arguments{
\item{n}{A numeric value specifying the number of days to simulate.}

\item{mu}{A numeric value giving the mean exposure.}

\item{sd}{A numeric value giving the standard deviation of the exposure
values around the trend line.}

\item{trend}{A character string that specifies the desired trend function.
Options are:
\itemize{
  \item{"cos1"}
  \item{"cos2"}
  \item{"cos3"}
  \item{"linear"}
  \item{"curvilinear"}
  \item{"cos1linear"}
  \item{"no trend"}
  \item{"custom"}
}}

\item{amp}{A numeric value specifying the amplitude of the seasonal trend.
Must be between 0 and 1.}

\item{...}{optional arguments to a custom trend function}
}
\value{
A data frame with the dates and daily exposure values from n days
}
\description{
Simulates a time series of continuous exposure values with or without a
seasonal trend.
}
\examples{
continuous_exposure(n = 5, mu = 100, sd = 10, trend = "cos1")

}

