% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{create_baseline}
\alias{create_baseline}
\title{Create a series of baseline outcomes}
\usage{
create_baseline(n, average_outcome, trend, custom_func = NULL, ...)
}
\description{
Create a series of baseline outcomes
}
\examples{
create_baseline(n = 5, average_outcome = 22, trend = "linear")
create_baseline(n = 5, average_outcome = NA, trend = NA,
                custom_func = "custom_baseline", outcome_type = "death")
}

