% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{create_lambda}
\alias{create_lambda}
\title{Create a series of mean outcome values}
\usage{
create_lambda(baseline, exposure, rr, cust_lambda_func = NULL, ...)
}
\arguments{
\item{baseline}{A numeric vector of baseline outcome values}

\item{exposure}{A numeric vector of exposure values}

\item{rr}{A numeric value specifying the relative risk}

\item{cust_lambda_func}{A character string specifying a user-made custom
function for relating baseline, relative risk, and exposure}

\item{...}{optional arguments for a custom lambda function}
}
\value{
A numeric vector of mean outcome values
}
\description{
This function relates exposure to baseline outcome values with the function
lambda = log(baseline) + log(relative risk)*exposure to create a series
of mean outcome values with or without incorporating a seasonal trend.  The
user may input a custom function to relate exposure, relative risk, and
baseline.
}
\examples{
base <- create_baseline(n = 10, average_baseline = 22, trend = "linear",
                        slope = .4)
exp <- sim_exposure(n = 5, central = 100, sd = 10, amp = .6,
                    exposure_type = "continuous")
create_lambda(baseline = base, exposure = exp$x, rr = 1.01)

}
