% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{custom_exposure}
\alias{custom_exposure}
\title{Pulls exposure series from Chicago NMMAPS data set}
\usage{
custom_exposure(n, central = NA, metric = "temp")
}
\description{
Pulls exposure series from Chicago NMMAPS data set
}
\examples{
custom_exposure(n = 5, metric = "temp")
}

