% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{eesim}
\alias{eesim}
\title{Wrapper to do Everything}
\usage{
eesim(n_reps, n, central, sd, exposure_type, exposure_trend, exposure_amp,
  average_outcome, outcome_trend = "no trend", outcome_amp, rr,
  start.date = "2000-01-01", cust_exp_func = NULL, cust_exp_args = NULL,
  cust_base_func = NULL, cust_lambda_func = NULL, cust_base_args = NULL,
  cust_lambda_args = NULL, model, df_year = NULL)
}
\description{
This function generates exposures and outcomes, fits models, and evaluates them for many repetitions
}
\examples{
eesim(n_reps = 3, n = 50, central = 100, sd = 10, exposure_type = "continuous",
exposure_trend = "cos3", exposure_amp = .6, average_outcome = 22, rr = 1.01,
model = "spline", df_year = 5)

}

