% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{fit_mods}
\alias{fit_mods}
\title{Fit models}
\usage{
fit_mods(outcome, model, df_year = 7)
}
\arguments{
\item{outcome}{A list of simulated data sets which each include columns called "x" and "outcome"}

\item{model}{A character string specifying model to be used.  Choices are "spline" and "casecrossover"}
}
\description{
Fit models
}

