% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{fit_mods}
\alias{fit_mods}
\title{Fit models}
\usage{
fit_mods(data, model = NULL, df_year = 7, custom_model = NULL,
  custom_model_args = list())
}
\arguments{
\item{data}{A list of simulated data sets. Each simulated dataset must include a
columns called "x" with daily exposure values and a column called "outcome" with
daily outcome values.}

\item{model}{A character string specifying model to be used. Choices are
"spline" and "casecrossover"}

\item{df_year}{A numeric value specifying the degrees of freedom per year}

\item{custom_model}{A character string with the object name of an R function
that defines the code that will be used to fit the model. See Details for more.}

\item{custom_model_args}{A list of arguments and their values for a custom
model.}
}
\value{
A data frame in which each row gives the results from the model-fitting function run
  on one of the simulated datasets input to the function as the \code{data} object. The returned
  data frame has one row per simulated dataset and the following columns:
  \itemize{
    \item{\code{Estimate}: The estimated \eqn{\beta} (log relative risk) as estimated by
      the model specified with \code{custom_model}.}
    \item{\code{Std.Error}: The standard error for the estimated \eqn{\beta}.}
    \item{\code{t.value}: The test statistic for a test of the null hypothesis \eqn{\beta = 0}.}
    \item{\code{p.value}: The p-value for a test of the null hypothesis \eqn{\beta = 0}.}
    \item{\code{lower_ci}: The lower value in the 95\% confidence interval estimated for
      \eqn{\beta}.}
    \item{\code{upper_ci}: The upper value in the 95\% confidence interval estimated for
      \eqn{\beta}.}
  }
}
\description{
Fits a specified model to each of the simulated datasets and returns a dataframe
summarizing results from fitting the model to each dataset, including the estimated
effect and the estimated standard error for that estimated effect.
}
\details{
The function specified by the \code{custom_model} argument should be
   a user-created function that inputs a data frame with columns named "x" for
   exposure values and "outcome" for outcome values. The function must output a data
   frame with columns called \code{Estimate}, \code{Std. Error}, \code{t value},
   \code{Pr(>|t|)}, \code{2.5\%}, and \code{97.5\%}. Note that these columns are the output
   from \code{summary} and \code{confint} for models fit using a \code{glm} call. For
   more details and examples, see the vignette for \code{eesim}.
}
\examples{
sims <- create_sims(n_reps = 10, n = 50, central = 100, sd = 10,
            exposure_type = "continuous", exposure_trend = "cos1",
            exposure_amp = .6, average_outcome = 22,
            outcome_trend = "no trend", outcome_amp = .6, rr = 1.01)
fit_mods(data = sims, model = "spline")

}
