% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{format_out}
\alias{format_out}
\title{Format output for custom model to use in eesim}
\usage{
format_out(mod)
}
\arguments{
\item{mod}{A model object from lm, glm, etc.}
}
\value{
Output with the correct column names for use as a custom model in eesim.
}
\description{
Format output for custom model to use in eesim
}
\examples{
dat <- data.frame(x=rnorm(1000, 0, 1), outcome = rnorm(1000, 5, 1))
lin_mod <- lm(outcome~x, data=dat)
format_out(lin_mod)

}
