% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_performance.R
\name{mean_beta}
\alias{mean_beta}
\title{Average Estimated Coefficient}
\usage{
mean_beta(df)
}
\arguments{
\item{df}{A data frame of replicated simulations which must include a column titled "est"}
}
\value{
A data frame with the mean estimated coefficient and mean estimated relative risk
}
\description{
This function gives the mean value of the \eqn{\hat{\beta}}s and the mean estimated
relative risk over the n simulations.
}
\examples{
mean_beta(df)

}

