% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_performance.R
\name{power_beta}
\alias{power_beta}
\title{Power}
\usage{
power_beta(df)
}
\arguments{
\item{df}{A data frame of replicated simulations which must include columns
titled "lower_ci" and "upper_ci"}
}
\value{
Power at the 5% significance level
}
\description{
This function gives the power of the test at a 5% siginificance level.
}
\examples{
power_beta(df)

}

