% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_performance.R
\name{power_beta}
\alias{power_beta}
\title{Power}
\usage{
power_beta(df)
}
\arguments{
\item{df}{A data frame of replicated simulations which must include columns
titled "lower_ci" and "upper_ci"}
}
\value{
Power at the 5% significance level
}
\description{
This function gives the power of the test at a 5% siginificance level.
}
\examples{
sims <- create_sims(n_reps = 10, n = 600, central = 100,
                    sd = 10, exposure_type = "continuous",
                    exposure_trend = "cos1",
                    exposure_amp = 0.6,
                    average_outcome = 20,
                    outcome_trend = "no trend",
                    rr = 1.01)
fits <- fit_mods(outcome = sims, model = "spline", df_year = 1)
power_beta(fits)

}

