% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exposure.R
\name{season_binexp}
\alias{season_binexp}
\title{Simulate binary exposure data with a seasonal trend}
\usage{
season_binexp(n, p, trend, start.date = "2000-01-01")
}
\arguments{
\item{n}{A numeric value giving the number of days to simulate.}

\item{p}{A numeric value giving the baseline probability of exposure for
a binary exposure.}
}
\value{
A numeric vector with simulated exposure of length n.
}
\description{
This function simulates binary exposure data with a seasonal trend.
}
\examples{
season_binexp(n = 5, p = 0.25)

}

