% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exposure.R
\name{season_contexp}
\alias{season_contexp}
\title{Simulate continuous exposure data}
\usage{
season_contexp(n, mu, sd, trend)
}
\arguments{
\item{n}{A numeric value giving the number of days to simulate.}

\item{mu}{A numeric vector giving the average of the exposure distribution.}

\item{sd}{A numeric vector giving the standard deviation of the exposure
distribution.}

\item{trend}{A character string that gives the trend function to use. Options
include ...}
}
\value{
A numeric vector with simulated exposure of length n.
}
\description{
This function simulates a time series of continuous exposure data.
}
\examples{
season_contexp(n = 5, mu = 100, sd = 10, trend = "cos1")

}

