% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{sim_exposure}
\alias{sim_exposure}
\title{Simulate random series of exposure values}
\usage{
sim_exposure(n, central, trend = "no trend", amp = 0.6,
  exposure_type = NA, start.date = "2001-01-01", custom_func = NULL, ...)
}
\description{
Simulate random series of exposure values
}
\examples{
sim_exposure(n = 5, central = 0.25, exposure_type = "binary", amp = .02)
sim_exposure(n = 5, central = 100, sd = 10, amp = .6,
             exposure_type = "continuous")
sim_exposure(n = 5, central = NA, custom_func = "custom_exposure",
                    metric = "temp")

}

