% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizable_functions.R
\name{sim_outcome}
\alias{sim_outcome}
\title{Simulate outcome}
\usage{
sim_outcome(exposure, average_outcome = NULL, trend = "no trend",
  slope = 1, amp = 0.6, rr = 1.01, start.date = "2000-01-01",
  cust_base_func = NULL, cust_lambda_func = NULL, cust_base_args = list(),
  cust_lambda_args = list(), cust_outdraw = NULL,
  cust_outdraw_args = list())
}
\arguments{
\item{exposure}{A numeric vector of exposure values}

\item{average_outcome}{A numeric value specifying the average outcome value.}

\item{trend}{outcome_trend A character string specifying the seasonal trend in
    health outcomes.  Options are Options are:
\itemize{
  \item{"cos1"}
  \item{"cos2"}
  \item{"cos3"}
  \item{"linear"}
  \item{"curvilinear"}
  \item{"cos1linear"}
  \item{"no trend"}
  \item{"custom"}
}}

\item{slope}{slope A numeric value specifying the slope of the trend, to be used
with trend = "linear" or trend = "cos1linear".}

\item{amp}{A numeric value specifying the amplitude of the seasonal trend.
Must be between 0 and 1.}

\item{rr}{A numeric value specifying the relative risk}

\item{start.date}{A date of the format "yyyy-mm-dd" from which to begin
simulating values}

\item{cust_base_func}{A character string specifying a user-made custom
function for baseline trend}

\item{cust_lambda_func}{A character string specifying a user-made custom
function for relating baseline, relative risk, and exposure}

\item{cust_base_args}{A list of arguments and their values used in the
user-specified custom baseline function}

\item{cust_lambda_args}{A list of arguments and their values used in the
user-specified custom lambda function}

\item{cust_outdraw}{A character string specifying a user-created function to
randomize the outcome values off of the baseline for outcome values. This
function must take inputs "n" and "lambda" and output a vector of outcome
values.}

\item{cust_outdraw_args}{A list of arguments besides "n" and "lamba" passed
to the user-created custom outcome draw function.}
}
\value{
A dataframe with a simulated time series, with columns for
   \code{date}, \code{x} (exposure), and \code{outcome}.
}
\description{
Simulate outcome
}
\examples{
exp <- sim_exposure(n = 5, central = 100, sd = 10, amp = .6,
                    exposure_type = "continuous")
sim_outcome(exposure = exp, average_outcome = 22, trend = "linear")

}
