% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{spline_mod}
\alias{spline_mod}
\title{Fit a generalized linear model}
\usage{
spline_mod(df, df_year = 7)
}
\arguments{
\item{df}{A data frame with exposure and outcome data called "x" and "outcome",
respectively}

\item{df_year}{A numeric value specifying the degrees of freedom per year}
}
\value{
Summary of the estimated log relative risk for a 1-unit increase in
   exposure
}
\description{
This function fits a GLM with splines to your data.
}
\examples{
exp <- sim_exposure(n = 500, central = 100, sd = 10, trend = "cos1",
                    amp = .6, exposure_type = "continuous")
out <- sim_outcome(exposure = exp, average_outcome = 22, rr = 1.01)
spline_mod(df = out)

}
